% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_frame.R
\name{clean_data_frame}
\alias{clean_data_frame}
\title{Cleans a data.frame by redacting PII information from character vector columns}
\usage{
clean_data_frame(df, spotters_list)
}
\arguments{
\item{df}{A data.frame to clean}

\item{spotters_list}{A list containing lists of 3 elements each:
\enumerate{
\item the redact function
\item hash_spotted value to pass or 0 to keep the default
\item the replace_with value or 0 to keep the default
}}
}
\value{
A deep copy of the cleaned data.frame.
}
\description{
Cleans a data.frame by redacting PII information from character vector columns
}
\examples{
df <- data.frame()
spotters <- list()
spotter_1 <- list(redact_email,TRUE,0)
spotters <- append(spotters,spotter_1)
df_cleaned <- clean_data_frame(df, spotters)
}
