% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_email.R
\name{redact_email}
\alias{redact_email}
\title{Redacts an email addresses from a given string}
\usage{
redact_email(string, hash_spotted = FALSE, replace_with = "EMAILADDRS")
}
\arguments{
\item{string}{A character vector with, at most, one element. The input string to redact email addresses from}

\item{hash_spotted}{When TRUE, the redaction of the email addresses will be a hash of the redacted (Default False)}

\item{replace_with}{A character vector with, at most, one element. When hash_spotted is FALSE, this character vector will be the replacement redacted email addresses.}
}
\value{
A character vector.
}
\description{
Redacts an email addresses from a given string
}
\examples{
x <- "my email address is foo@gaga.com"
redact_email(x)
}
