% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_frame.R
\name{redact_it}
\alias{redact_it}
\title{Cleans a data.frame by redacting PII information from character vector columns}
\usage{
redact_it(string, spotters_list)
}
\arguments{
\item{string}{A string to redact}

\item{spotters_list}{A list containing lists of 3 elements each:
\enumerate{
\item the redact function
\item hash_spotted value to pass or 0 to keep the default
\item the replace_with value or 0 to keep the default
}}
}
\value{
A redacted string
}
\description{
Cleans a data.frame by redacting PII information from character vector columns
}
